@@ -1,14 +1,14 @@
 package com.example.android.vacuumfitness.ui;
 
 
-import android.arch.lifecycle.Observer;
-import android.arch.lifecycle.ViewModelProviders;
+import androidx.lifecycle.Observer;
+import androidx.lifecycle.ViewModelProvider;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.appcompat.widget.Toolbar;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -89,8 +89,8 @@ public class AllExercisesFragment extends Fragment implements ChooseExerciseAdap
     }
 
     private void setupViewModel() {
-        ExerciseListViewModel viewModel = ViewModelProviders.of(this).get(ExerciseListViewModel.class);
-        viewModel.getExercises().observe(this, new Observer<List<Exercise>>() {
+        ExerciseListViewModel viewModel = new ViewModelProvider(this).get(ExerciseListViewModel.class);
+        viewModel.getExercises().observe(getViewLifecycleOwner(), new Observer<List<Exercise>>() {
             @Override
             public void onChanged(@Nullable List<Exercise> exerciseList) {
                 exerciseAdapter.setExercises(exerciseList);
