@@ -2,18 +2,18 @@ package com.example.android.vacuumfitness.ui;
 
 
 import android.app.AlertDialog;
-import android.arch.lifecycle.Observer;
-import android.arch.lifecycle.ViewModelProviders;
+import androidx.lifecycle.Observer;
+import androidx.lifecycle.ViewModelProvider;
 import android.content.DialogInterface;
 import android.graphics.drawable.GradientDrawable;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
+import androidx.annotation.Nullable;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentTransaction;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.appcompat.widget.Toolbar;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -93,8 +93,8 @@ public class CustomizeTrainingFragment extends Fragment implements CustomTrainin
 
     //Setup CustomTrainingViewModel
     private void setupCustomTrainingViewModel(){
-        CustomTrainingViewModel viewModel = ViewModelProviders.of(this).get(CustomTrainingViewModel.class);
-        viewModel.getTrainings().observe(this, new Observer<List<Training>>() {
+        CustomTrainingViewModel viewModel = new ViewModelProvider(this).get(CustomTrainingViewModel.class);
+        viewModel.getTrainings().observe(getViewLifecycleOwner(), new Observer<List<Training>>() {
             @Override
             public void onChanged(@Nullable List<Training> trainings) {
                 if(trainings.isEmpty()){
