@@ -3,18 +3,17 @@ package com.example.android.vacuumfitness.ui;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.support.design.widget.CollapsingToolbarLayout;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v7.app.AppCompatActivity;
+import com.google.android.material.appbar.CollapsingToolbarLayout;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.widget.Toolbar;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import androidx.work.Constraints;
 import androidx.work.NetworkType;
-import androidx.work.OneTimeWorkRequest;
 import androidx.work.PeriodicWorkRequest;
 import androidx.work.WorkManager;
 import com.example.android.vacuumfitness.R;
@@ -23,7 +22,6 @@ import com.example.android.vacuumfitness.utils.AppExecutors;
 import com.example.android.vacuumfitness.utils.KeyUtils;
 import com.example.android.vacuumfitness.utils.SharedPrefsUtils;
 import com.example.android.vacuumfitness.worker.ScheduleSingleNotificationWorker;
-import com.example.android.vacuumfitness.worker.SendNotificationWorker;
 import com.example.android.vacuumfitness.worker.UpdateMotivatorWorker;
 import com.google.firebase.analytics.FirebaseAnalytics;
 import java.util.concurrent.TimeUnit;
@@ -36,7 +34,6 @@ public class MainActivity extends AppCompatActivity implements FragmentManager.O
     private static String NOTIFICATION_WORK_TAG = "notification-work-tag";
 
     @BindView(R.id.toolbar) Toolbar toolbar;
-    @BindView(R.id.collapsing_toolbar) CollapsingToolbarLayout collapsingToolbar;
 
     public static SharedPreferences sharedPreferences;
     public static Context mContext;
