@@ -1,12 +1,12 @@
 package com.example.android.vacuumfitness.ui;
 
 
-import android.arch.lifecycle.Observer;
-import android.arch.lifecycle.ViewModelProviders;
+import androidx.lifecycle.Observer;
+import androidx.lifecycle.ViewModelProvider;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -132,8 +132,8 @@ public class PreparationFragment extends Fragment {
 
     //Setup CustomTrainingViewModel
     private void setupTrainingViewModel(){
-        CustomTrainingViewModel viewModel = ViewModelProviders.of(this).get(CustomTrainingViewModel.class);
-        viewModel.getTrainings().observe(this, new Observer<List<Training>>() {
+        CustomTrainingViewModel viewModel = new ViewModelProvider(this).get(CustomTrainingViewModel.class);
+        viewModel.getTrainings().observe(getViewLifecycleOwner(), new Observer<List<Training>>() {
             @Override
             public void onChanged(@Nullable List<Training> trainings) {
                 mTrainingList = SpinnerUtils.populateTrainingSpinnerItems(getActivity(), trainingSpinner, trainings);
@@ -149,8 +149,8 @@ public class PreparationFragment extends Fragment {
     }
 
     private void setupPlaylistViewModel(){
-        PlaylistViewModel viewModel = ViewModelProviders.of(this).get(PlaylistViewModel.class);
-        viewModel.getPlaylists().observe(this, new Observer<List<Playlist>>() {
+        PlaylistViewModel viewModel = new ViewModelProvider(this).get(PlaylistViewModel.class);
+        viewModel.getPlaylists().observe(getViewLifecycleOwner(), new Observer<List<Playlist>>() {
             @Override
             public void onChanged(@Nullable List<Playlist> playlists) {
                 mPlaylistList = SpinnerUtils.populateMusicSpinnerItems(getActivity(), musicSpinner, playlists);
