@@ -1,7 +1,7 @@
 package com.example.android.vacuumfitness.worker;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.Log;
 
 import androidx.work.Worker;
