@@ -1,9 +1,9 @@
 package com.example.android.vacuumfitness.database;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.Query;
+import androidx.lifecycle.LiveData;
+import androidx.room.Dao;
+import androidx.room.Insert;
+import androidx.room.Query;
 import com.example.android.vacuumfitness.model.Exercise;
 import java.util.List;
 
