@@ -1,6 +1,6 @@
 package com.example.android.vacuumfitness.utils;
 
-import android.arch.persistence.room.TypeConverter;
+import androidx.room.TypeConverter;
 
 import com.example.android.vacuumfitness.model.Exercise;
 import com.example.android.vacuumfitness.model.Song;
