@@ -1,7 +1,7 @@
 package com.example.android.vacuumfitness.viewmodel;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.lifecycle.ViewModel;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.ViewModel;
 import android.content.Context;
 import android.util.Log;
 
